+++ Kexa/services/addOn/awsGathering.service.ts
@@ -176,7 +176,7 @@ async function ec2SGListing(client: EC2, region: string): Promise<any> {
         logger.debug(region + " - ec2SGListing Done");
         return jsonData;
     } catch (err) {
-        logger.error("Error in ec2SGListing: ", err);
+        logger.debug("Error in ec2SGListing: ", err);
         return null;
     }
 }
@@ -189,7 +189,7 @@ async function ec2VolumesListing(client: EC2, region: string): Promise<any> {
         logger.debug(region, " - ec2VolumesListing Done");
         return jsonData;
     } catch (err) {
-        logger.error("Error in ec2VolumesListing: ", err);
+        logger.debug("Error in ec2VolumesListing: ", err);
         return null;
     }
 }
@@ -202,7 +202,7 @@ async function ec2InstancesListing(client: EC2, region: string): Promise<Array<E
         logger.debug(region + " - ec2InstancesListing Done");
         return jsonData;
     } catch (err) {
-        logger.error("Error in ec2InstancesListing: ", err);
+        logger.debug("Error in ec2InstancesListing: ", err);
         return null;
     }
 }
@@ -215,7 +215,7 @@ async function rdsInstancesListing(client: RDS, region: string): Promise<any> {
         logger.debug(region + " - rdsInstancesListing Done");
         return jsonData;
     } catch (err) {
-        logger.error("Error in rdsInstancesListing: ", err);
+        logger.debug("Error in rdsInstancesListing: ", err);
         return null;
     }
 }
@@ -228,7 +228,7 @@ async function resourceGroupsListing(client: ResourceGroups, region: string): Pr
         logger.debug(region + " - Ressource Group Done");
         return jsonData;
     } catch (err) {
-        logger.error("Error in Ressource Group Listing: ", err);
+        logger.debug("Error in Ressource Group Listing: ", err);
         return null;
     }
 }
@@ -249,7 +249,7 @@ async function tagsValueListing(client: ResourceGroupsTaggingAPI, region: string
         logger.debug(region + " - Tags Done");
         return jsonDataKeys;
     } catch (err) {
-        logger.error("Error in Tags Value Listing: ", err);
+        logger.debug("Error in Tags Value Listing: ", err);
         return null;
     }
 }
@@ -262,7 +262,7 @@ async function s3BucketsListing(client: S3, region: string): Promise<Array<S3> |
         logger.debug(region + " - s3BucketsListing Done");
         return jsonData;
     } catch (err) {
-        logger.error("Error in s3BucketsListing: ", err);
+        logger.debug("Error in s3BucketsListing: ", err);
         return null;
     }
 }
@@ -275,7 +275,7 @@ async function ecsClusterListing(client: ECS, region: string): Promise<any> {
         logger.debug(region + " - ECS Done");
         return jsonData;
     } catch (err) {
-        logger.error("Error in ECS Listing: ", err);
+        logger.debug("Error in ECS Listing: ", err);
         return null;
     }
 }
@@ -288,7 +288,7 @@ async function ecrImagesListing(client: ECR, region: string): Promise<any> {
         logger.debug(region + " - ECR Done");
         return jsonData;
     } catch (err) {
-        logger.error("Error in ECR Listing: ", err);
+        logger.debug("Error in ECR Listing: ", err);
         return null;
     }
 }