+++ Kexa/services/addOn/azureGathering.service.ts
@@ -12,6 +12,10 @@
     *     - virtualNetwork
     *     - networkInterfaces
     *     - aks
+    *     - mlWorkspace
+    *     - mlJobs
+    *     - mlComputes
+    *     - mlSchedule
 */
 
 import { 
@@ -52,6 +56,10 @@ export async function collectData(azureConfig:AzureConfig[]): Promise<AzureResou
             "virtualNetwork": null,
             "networkInterfaces": null,
             "aks": null,
+            "mlWorkspaces": null,
+            "mlJobs": null,
+            "mlComputes": null,
+            "mlSchedules": null,
         } as AzureResources;
         logger.debug("config: ");
         logger.debug(JSON.stringify(config));
@@ -91,20 +99,25 @@ export async function collectData(azureConfig:AzureConfig[]): Promise<AzureResou
                     virtualNetworksListing(networkClient),
                     aksListing(credential, subscriptionId),
                     ipListing(networkClient),
+                    mlListing(credential, subscriptionId),
                     //getSPKeyInformation(credential, subscriptionId)
                 ];
                 
-                const [nsgList, vmList, rgList, diskList, virtualNetworkList, aksList, ipList] = await Promise.all(promises); //, SPList
+                const [nsgList, vmList, rgList, diskList, virtualNetworkList, aksList, ipList, mlList] = await Promise.all(promises); //, SPList
                 context?.log("- listing cloud resources done -");
                 logger.info("- listing cloud resources done -");
                 azureResource = {
-                    "vm": [...azureResource["vm"]??[], ...vmList],
-                    "rg": [...azureResource["rg"]??[], ...rgList],
-                    "disk": [...azureResource["disk"]??[], ...diskList],
-                    "nsg": [...azureResource["nsg"]??[], ...nsgList],
-                    "virtualNetwork": [...azureResource["virtualNetwork"]??[], ...virtualNetworkList],
-                    "aks": [...azureResource["aks"]??[], ...aksList],
-                    "ip": [...azureResource["ip"]??[], ...ipList],
+                    "vm": [...azureResource["vm"]??[], ...vmList??[]],
+                    "rg": [...azureResource["rg"]??[], ...rgList??[]],
+                    "disk": [...azureResource["disk"]??[], ...diskList??[]],
+                    "nsg": [...azureResource["nsg"]??[], ...nsgList??[]],
+                    "virtualNetwork": [...azureResource["virtualNetwork"]??[], ...virtualNetworkList??[]],
+                    "aks": [...azureResource["aks"]??[], ...aksList??[]],
+                    "ip": [...azureResource["ip"]??[], ...ipList??[]],
+                    "mlWorkspaces": [...azureResource["mlWorkspaces"]??[], ...mlList.workspaces??[]],
+                    "mlJobs": [...azureResource["mlJobs"]??[], ...mlList.jobs??[]],
+                    "mlComputes": [...azureResource["mlComputes"]??[], ...mlList.computes??[]],
+                    "mlSchedules": [...azureResource["mlSchedules"]??[], ...mlList.schedule??[]],
                     //"sp": [...azureResource["sp"]??[], ...SPList],
                 } as AzureResources;
             }
@@ -129,7 +142,7 @@ export async function getSPKeyInformation(credential: DefaultAzureCredential, su
         }
         return resultList;
     } catch (err) {
-        logger.error("error in getSPKeyInformation:"+err);
+        logger.debug("error in getSPKeyInformation:"+err);
         return null;
     }
 }
@@ -144,7 +157,7 @@ export async function ipListing(client:NetworkManagementClient): Promise<Array<a
         }
         return resultList;
     }catch(e){
-        logger.error("error in ipListing:"+e);
+        logger.debug("error in ipListing:"+e);
         return null;
     }
 }
@@ -160,7 +173,7 @@ export async function aksListing(credential: DefaultAzureCredential, subscriptio
         }
         return resArray;
     }catch(e){
-        logger.error("error in aksListing:"+e);
+        logger.debug("error in aksListing:"+e);
         return null;
     }
 }
@@ -176,7 +189,7 @@ export async function networkSecurityGroupListing(client:NetworkManagementClient
         logger.info("ended networkSecurityGroupListing");
         return resultList;        
     } catch (err) {
-        logger.error("error in networkSecurityGroupListing:"+err);
+        logger.debug("error in networkSecurityGroupListing:"+err);
         return null;
     }
 }
@@ -192,7 +205,7 @@ export async function virtualNetworksListing(client:NetworkManagementClient): Pr
 
         return resultList;
     } catch (err) {
-        logger.error("error in virtualNetworksListing:"+err);
+        logger.debug("error in virtualNetworksListing:"+err);
         return null;
     }
 }
@@ -207,7 +220,7 @@ export async function networkInterfacesListing(client:NetworkManagementClient): 
         }
         return resultList;
     } catch (err) {
-        logger.error("error in networkInterfacesListing:"+err);
+        logger.debug("error in networkInterfacesListing:"+err);
         return null;
     }
 }
@@ -222,7 +235,7 @@ export async function disksListing(client:ComputeManagementClient): Promise<Arra
         }
         return resultList;
     } catch (err) {
-        logger.error("error in disksListing:"+err);
+        logger.debug("error in disksListing:"+err);
         return null;
     }    
 }
@@ -237,12 +250,11 @@ export async function virtualMachinesListing(client:ComputeManagementClient): Pr
         }
         return resultList;
     }catch (err) {
-        logger.error("error in virtualMachinesListing:"+err);
+        logger.debug("error in virtualMachinesListing:"+err);
         return null;
     } 
 }
 
-//resourceGroups.list
 export async function resourceGroupListing(client:ResourceManagementClient): Promise<Array<ResourceGroup>|null> {
     logger.info("starting resourceGroupListing");
     try {
@@ -252,13 +264,10 @@ export async function resourceGroupListing(client:ResourceManagementClient): Pro
         }
         return resultList;
     }catch (err) {
-        logger.error("error in resourceGroupListing:"+err);
+        logger.debug("error in resourceGroupListing:"+err);
         return null;
     }     
 }
-
-///////////////////////////////////////////////////////////////////////////////////////////////////////
-
 
 export async function networkSecurityGroup_analyse(nsgList: Array<NetworkSecurityGroup>): Promise<Array<ckiNetworkSecurityClass.CkiNetworkSecurityGroupClass>|null> {
     try {
@@ -274,7 +283,97 @@ export async function networkSecurityGroup_analyse(nsgList: Array<NetworkSecurit
         }
         return resultList;
     }catch (e) {
-        logger.error("error"+e);
+        logger.debug("error"+e);
         return null;
     }  
+}
+
+
+import { AzureMachineLearningWorkspaces } from "@azure/arm-machinelearning";
+export async function mlListing(credential: DefaultAzureCredential, subscriptionId: string): Promise<any> {
+    logger.info("starting mlListing");
+    try{
+        const client = new AzureMachineLearningWorkspaces(credential, subscriptionId);
+        const result = {
+            "workspaces": new Array(),
+            "jobs": new Array(),
+            "computes": new Array(),
+            "schedule": new Array(),
+        }
+        for await (let item of client.workspaces.listBySubscription()) {
+            result.workspaces = [...result.workspaces??[], item];
+            let resourceGroupName = item?.id?.split("/")[4] ?? "";
+            let workspaceName = item?.name ?? "";
+            const promises = [
+                jobsListing(client, resourceGroupName, workspaceName),
+                computeOperationsListing(client, resourceGroupName, workspaceName),
+                schedulesListing(client, resourceGroupName, workspaceName),
+            ];
+            const [jobsList, computeOperationsList, schedulesList] = await Promise.all(promises);
+            logger.error("jobsList: "+JSON.stringify(jobsList));
+            logger.error("computeOperationsList: "+JSON.stringify(computeOperationsList));
+            logger.error("schedulesList: "+JSON.stringify(schedulesList));
+            result.jobs = [...result.jobs??[], ...jobsList];
+            result.computes = [...result.computes??[], ...computeOperationsList];
+            result.schedule = [...result.schedule??[], ...schedulesList];
+            logger.error("RESULT0: ");
+            logger.error("RESULT0: "+JSON.stringify(result));
+        }
+        logger.error("RESULT1: "+JSON.stringify(result));
+        return result;
+    }catch(e){
+        logger.debug("error in mlListing:"+e);
+        return null;
+    }
+}
+
+export async function jobsListing(client: AzureMachineLearningWorkspaces, resourceGroupName: string, workspaceName: string): Promise<any[]> {
+    //logger.info("starting jobsListing");
+    try{
+        const resArray = new Array();
+        for await (let item of client.jobs.list(resourceGroupName, workspaceName)) {
+            let result:any = item;
+            result.workspace = workspaceName;
+            result.resourceGroupName = resourceGroupName;
+            resArray.push(result);
+        }
+        return resArray;
+    }catch(e){
+        logger.debug("error in jobsListing:"+e);
+        return [];
+    }
+}
+
+export async function computeOperationsListing(client: AzureMachineLearningWorkspaces, resourceGroupName: string, workspaceName: string): Promise<any[]> {
+    //logger.info("starting computeOperationsListing");
+    try{
+        const resArray = new Array();
+        for await (let item of client.computeOperations.list(resourceGroupName, workspaceName)) {
+            let result:any = item;
+            result.workspace = workspaceName;
+            result.resourceGroupName = resourceGroupName;
+            resArray.push(item);
+        }
+        return resArray;
+    }catch(e){
+        logger.debug("error in computeOperationsListing:"+e);
+        return [];
+    }
+}
+
+export async function schedulesListing(client: AzureMachineLearningWorkspaces, resourceGroupName: string, workspaceName: string): Promise<any[]> {
+    //logger.info("starting schedulesListing");
+    try{
+        const resArray = new Array();
+        for await (let item of client.schedules.list(resourceGroupName, workspaceName)) {
+            let result:any = item;
+            result.workspace = workspaceName;
+            result.resourceGroupName = resourceGroupName;
+            resArray.push(item);
+        }
+        return resArray;
+    }catch(e){
+        logger.debug("error in schedulesListing:"+e);
+        return [];
+    }
 }