+++ Kexa/services/addOn/gcpGathering.service.ts
@@ -224,7 +224,7 @@ function compareUserAndValidRegions(userRegions: Array<any>, validRegions: Array
         if (validRegions.includes(userRegions[i]))
             continue;
         else {
-            logger.error("GCP - Config '" + gcpConfig.indexOf(config) + "' Skipped - Regions '" + userRegions[i] + "' is not a valid GCP region.");
+            logger.warn("GCP - Config '" + gcpConfig.indexOf(config) + "' Skipped - Regions '" + userRegions[i] + "' is not a valid GCP region.");
             return (false);
         }
     }
@@ -247,7 +247,7 @@ async function retrieveAllRegions(projectId: number, regionsList: Array<string>)
             regionsList.push(response.name);
         }
     } catch (e) {
-        logger.error("GCP : Error while retrieving all regions")
+        logger.debug("GCP : Error while retrieving all regions")
     }
 }
 
@@ -308,7 +308,7 @@ async function listTasks(projectId: number, regionsList: Array<string>): Promise
         const tasksClient = new CloudTasksClient();
         jsonData = await executeAllRegions(projectId, tasksClient.listQueuesAsync, tasksClient, regionsList, true);
     } catch (e) {
-        logger.error("Error while retrieving GCP Tasks Queues");
+        logger.debug("Error while retrieving GCP Tasks Queues");
     }
     return jsonData ?? null;
 }
@@ -383,7 +383,7 @@ async function listAllBucket(): Promise<Array<any>|null> {
 
         }
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Buckets Listing Done");
     return jsonReturn ?? null;
@@ -406,7 +406,7 @@ async function listAllClusters(): Promise<Array<any>|null> {
         const [response] = await cnt.listClusters(request);
         jsonData = JSON.parse(JSON.stringify(response));
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Clusters Listing Done");
     return jsonData ?? null;
@@ -423,7 +423,7 @@ async function listAllProject(): Promise<Array<any>|null> {
             jsonData = JSON.parse(JSON.stringify(project));
         }
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Projects Listing Done");
     return jsonData ?? null;
@@ -442,7 +442,7 @@ async function getBillingAccount(projectId: any): Promise<Array<any>|null> {
             jsonData = JSON.parse(JSON.stringify(account));
         }
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Billing Accounts Listing Done");
     return jsonData ?? null;
@@ -460,7 +460,7 @@ async function listWorkflows(projectId: any): Promise<Array<any>|null> {
             jsonData = JSON.parse(JSON.stringify(workflow));
         }
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Workflows Listing Done");
     return jsonData ?? null;
@@ -477,7 +477,7 @@ async function listWebSecurityConfig(projectId: any): Promise<Array<any>|null> {
         jsonData = JSON.parse(JSON.stringify(stats));
     }
     catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Web Security Configs Listing Done");
     return jsonData ?? null;
@@ -490,7 +490,7 @@ async function listVpcConnectors(projectId: any, regionsList: Array<string>): Pr
         const client = new VpcAccessServiceClient();
         jsonData = await executeAllRegions(projectId, client.listConnectors, client, regionsList, false )
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP VPC Connectors Listing Done");
     return jsonData ?? null;
@@ -512,7 +512,7 @@ async function listVMWareEngine(projectId: any): Promise<Array<any>|null>  {
             jsonData = JSON.parse(JSON.stringify(response));
         }
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP VMWare Engine Listing Done");
     return jsonData ?? null;
@@ -527,7 +527,7 @@ async function listNamespaces(projectId: any, regionsList: Array<string>): Promi
         jsonData = await executeAllRegions(projectId, registrationServiceClient.listNamespaces, registrationServiceClient,
             regionsList, true);
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Namespaces Listing Done");
     return jsonData ?? null;
@@ -546,7 +546,7 @@ export async function listSecrets(projectId: any): Promise<Array<any>|null> {
             jsonData = JSON.parse(JSON.stringify(response));
         }
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Secrets Listing Done");
     return jsonData ?? null;
@@ -563,7 +563,7 @@ async function listConnectivityTests(projectId: any): Promise<Array<any>|null>  
         });
         jsonData = JSON.parse(JSON.stringify(tests));
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Connectivity Tests Listing Done");
     return jsonData ?? null;
@@ -581,7 +581,7 @@ async function listResourceSettings(projectId: any): Promise<Array<any>|null> {
         });
         jsonData = JSON.parse(JSON.stringify(settings));
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Resources Settings Listing Done");
     return jsonData ?? null;
@@ -595,7 +595,7 @@ async function listRedisInstances(projectId: any, regionsList: Array<string>): P
         const client = new CloudRedisClient();
         jsonData = await executeAllRegions(projectId, client.listInstances, client, regionsList, false);
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Redis Instances Listing Done");
     return jsonData ?? null;
@@ -612,7 +612,7 @@ async function listOSConfig(projectId: any): Promise<Array<any>|null> {
         });
         jsonData = JSON.parse(JSON.stringify(deployments));
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP OS Config Listing Done");
     return jsonData ?? null;
@@ -629,7 +629,7 @@ async function listOrgPolicyContraints(projectId: any): Promise<Array<any>|null>
         });
         jsonData = JSON.parse(JSON.stringify(constraints));
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP OrgPolicy Contraints Listing Done");
     return jsonData ?? null;
@@ -643,7 +643,7 @@ async function listOrchestrationAirflow(projectId: any, regionsList: Array<strin
         const client = new ImageVersionsClient();
         jsonData = await executeAllRegions(projectId, client.listImageVersions, client, regionsList, false);
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP OrgPolicy Contraints Listing Done");
     return jsonData ?? null;
@@ -662,7 +662,7 @@ async function listNotebookInstances(projectId: any): Promise<Array<any> | null>
             jsonData = JSON.parse(JSON.stringify(instance));
         }
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Notebook Instances Listing Done");
     return jsonData ?? null;
@@ -680,7 +680,7 @@ async function listDashboards(projectId: any): Promise<Array<any> | null> {
             jsonData = JSON.parse(JSON.stringify(dashboard));
         }
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Dashboards Listing Done");
     return jsonData ?? null;
@@ -697,7 +697,7 @@ async function listIdentitiesDomain(projectId: any): Promise<Array<any> | null> 
         });
         jsonData = JSON.parse(JSON.stringify(domains));
     } catch (e) {
-        logger.error(e)
+        logger.debug(e)
     }
     logger.info("GCP Identities Domains Listing Done");
     return jsonData ?? null;
@@ -716,7 +716,7 @@ async function listLineageProcesses(projectId: any): Promise<Array<any> | null> 
             jsonData = JSON.parse(JSON.stringify(response));
         }
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Lineage Processes Listing Done");
     return jsonData ?? null;
@@ -736,7 +736,7 @@ async function listKMSCryptoKeys(projectId: any): Promise<Array<any> | null> {
         });
         jsonData = JSON.parse(JSON.stringify(response));
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP KMS Crypto Keys Listing Done");
     return jsonData ?? null;
@@ -756,7 +756,7 @@ async function listKMSKeyRings(projectId: any): Promise<Array<any> | null> {
             jsonData = JSON.parse(JSON.stringify(keyRing));
         }
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP KMS Key Rings Listing Done");
     return jsonData ?? null;
@@ -773,7 +773,7 @@ async function listDomainsRegistration(projectId: any): Promise<Array<any> | nul
         });
         jsonData = JSON.parse(JSON.stringify(registrations));
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Domains Resgitration Listing Done");
     return jsonData ?? null;
@@ -790,7 +790,7 @@ async function listDnsZones(projectId: any): Promise<Array<any> | null> {
             jsonData = JSON.parse(JSON.stringify(zone));
         }
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP DNS Zones Listing Done");
     return jsonData ?? null;
@@ -804,7 +804,7 @@ async function listDeliveryPipelines(projectId: any, regionsList: Array<string>)
         const deployClient = new CloudDeployClient();
         jsonData = await executeAllRegions(projectId, deployClient.listDeliveryPipelinesAsync, deployClient, regionsList, true);
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Delivery Pipelines Listing Done");
     return jsonData ?? null;
@@ -823,7 +823,7 @@ async function listCertificates(projectId: any): Promise<Array<any> | null> {
             jsonData = JSON.parse(JSON.stringify(response));
         }
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Certificates Listing Done");
     return jsonData ?? null;
@@ -837,7 +837,7 @@ async function listBatchJobs(projectId: any, regionsList: Array<string>): Promis
         const batchClient = new BatchServiceClient();
         jsonData = await executeAllRegions(projectId, batchClient.listJobsAsync, batchClient, regionsList, true);
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Batch Jobs Listing Done");
     return jsonData ?? null;
@@ -860,7 +860,7 @@ async function listWorkloads(projectId: any): Promise<Array<any> | null> {
             jsonData = JSON.parse(JSON.stringify(workload));
         }*/
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Workloads Listing Done");
     return jsonData ?? null;
@@ -874,7 +874,7 @@ async function listArtifactsRepositories(projectId: any, regionsList: Array<stri
         const client = new ArtifactRegistryClient();
         jsonData = await executeAllRegions(projectId, client.listRepositories, client, regionsList, false);
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP Artifacts Repositories Listing Done");
     return jsonData ?? null;
@@ -888,7 +888,7 @@ async function listAppGateways(projectId: any, regionsList: Array<string>): Prom
         const appgatewaysClient = new AppGatewaysServiceClient();
         jsonData = await executeAllRegions(projectId, appgatewaysClient.listAppGatewaysAsync, appgatewaysClient, regionsList, true);
     } catch (e) {
-        logger.error(e);
+        logger.debug(e);
     }
     logger.info("GCP App Gateways Listing Done");
     return jsonData ?? null;