+++ CHANGELOG.md
@@ -21,3 +21,274 @@ Kexa/services/updateCapability.service.ts was changed
 package.json was changed
 
 
+## 1.8.0-SNAPSHOT.53.56373d8
+
+### Files added: 0
+
+### Files changed: 10
+
+Kexa/rules/Deployement.yaml was changed
+
+Kexa/rules/Economy.yaml was changed
+
+Kexa/rules/HTTPRules.yaml was changed
+
+Kexa/rules/OperationalExcellence.yaml was changed
+
+Kexa/rules/Performance.yaml was changed
+
+Kexa/rules/PostDeployement.yaml was changed
+
+Kexa/rules/PreDeployement.yaml was changed
+
+Kexa/rules/Security.yaml was changed
+
+Kexa/rules/driveRules.yaml was changed
+
+Kexa/rules/rules-testing.yaml was changed
+
+
+## 1.8.0-SNAPSHOT.55.c3f6d1b
+
+### Files added: 0
+
+### Files changed: 1
+
+README.md was changed
+
+
+## 1.8.0-SNAPSHOT.58.885a664
+
+### Files added: 0
+
+### Files changed: 6
+
+Kexa/rules/Economy.yaml was changed
+
+Kexa/rules/OperationalExcellence.yaml was changed
+
+Kexa/rules/Performance.yaml was changed
+
+Kexa/rules/Security.yaml was changed
+
+Kexa/rules/rules-testing.yaml was changed
+
+README.md was changed
+
+
+## 1.8.0-SNAPSHOT.61.b5ef343
+
+### Files added: 1
+
+images/schema-engine.png was added
+
+### Files changed: 2
+
+images/schema-engine.png was changed
+
+documentation/Documentation-Kexa.md was changed
+
+
+## 1.8.0-SNAPSHOT.67.e122e26
+
+### Files added: 0
+
+### Files changed: 14
+
+.gitignore was changed
+
+Kexa/emails/emails.ts was changed
+
+Kexa/helpers/files.ts was changed
+
+Kexa/main.ts was changed
+
+Kexa/models/resultScan.models.ts was changed
+
+Kexa/models/settingFile/rules.models.ts was changed
+
+Kexa/rules/Deployement.yaml was changed
+
+Kexa/services/alerte.service.ts was changed
+
+Kexa/services/analyse.service.ts was changed
+
+Kexa/services/display.service.ts was changed
+
+README.md was changed
+
+config/default.json was changed
+
+documentation/Documentation-Kexa.md was changed
+
+package.json was changed
+
+
+## 1.8.0-SNAPSHOT.69.12ed96c
+
+### Files added: 0
+
+### Files changed: 8
+
+Kexa/services/addOn/awsGathering.service.ts was changed
+
+Kexa/services/addOn/azureGathering.service.ts was changed
+
+Kexa/services/addOn/gcpGathering.service.ts was changed
+
+Kexa/services/addOn/githubGathering.service.ts was changed
+
+Kexa/services/addOn/googleDriveGathering.service.ts was changed
+
+Kexa/services/addOn/googleWorkspaceGathering.service.ts was changed
+
+Kexa/services/addOn/kubernetesGathering.service.ts was changed
+
+Kexa/services/addOn/o365Gathering.service.ts was changed
+
+
+## 1.8.0-SNAPSHOT.76.2e20887
+
+### Files added: 0
+
+### Files changed: 1
+
+Kexa/services/addOn/o365Gathering.service.ts was changed
+
+
+## 1.8.0-SNAPSHOT.80.a324899
+
+### Files added: 0
+
+### Files changed: 2
+
+README.md was changed
+
+documentation/Documentation-Kexa.md was changed
+
+
+## 1.8.0-SNAPSHOT.91.b5c4d86
+
+### Files added: 0
+
+### Files changed: 3
+
+Kexa/services/analyse.service.ts was changed
+
+config/env/office365.json was changed
+
+documentation/Documentation-Kexa.md was changed
+
+
+## 1.8.0-SNAPSHOT.93.aa30113
+
+### Files added: 1
+
+config/default.json was added
+
+### Files changed: 1
+
+config/default.json was changed
+
+
+## 1.8.0-SNAPSHOT.98.874b280
+
+### Files added: 0
+
+### Files changed: 5
+
+Kexa/models/azure/resource.models.ts was changed
+
+Kexa/rules/OperationalExcellence.yaml was changed
+
+Kexa/services/addOn/azureGathering.service.ts was changed
+
+package-lock.json was changed
+
+package.json was changed
+
+
+## 1.8.0-SNAPSHOT.100.29206c4
+
+### Files added: 0
+
+### Files changed: 1
+
+README.md was changed
+
+
+## 1.8.0-SNAPSHOT.102.b8694e9
+
+### Files added: 0
+
+### Files changed: 3
+
+.gitignore was changed
+
+documentation/Documentation-Kexa.md was changed
+
+
+## 1.8.0-SNAPSHOT.109.96001d8
+
+### Files added: 10
+
+Kexa/rules/rulesByProvider/awsSetRules.yaml was added
+
+Kexa/rules/rulesByProvider/azureSetRules.yaml was added
+
+Kexa/rules/rulesByProvider/gcpSetRules.yaml was added
+
+Kexa/rules/rulesByProvider/githubSetRules.yaml was added
+
+Kexa/rules/rulesByProvider/googleDriveSetRules.yaml was added
+
+Kexa/rules/rulesByProvider/googleWorkspaceSetRules.yaml was added
+
+Kexa/rules/rulesByProvider/httpSetRules.yaml was added
+
+Kexa/rules/rulesByProvider/kubernetesSetRules.yaml was added
+
+Kexa/rules/rulesByProvider/o365SetRules.yaml was added
+
+initKexa.ps1 was added
+
+### Files changed: 14
+
+Kexa/rules/rulesByProvider/awsSetRules.yaml was changed
+
+Kexa/rules/rulesByProvider/azureSetRules.yaml was changed
+
+Kexa/rules/rulesByProvider/gcpSetRules.yaml was changed
+
+Kexa/rules/rulesByProvider/githubSetRules.yaml was changed
+
+Kexa/rules/rulesByProvider/googleDriveSetRules.yaml was changed
+
+Kexa/rules/rulesByProvider/googleWorkspaceSetRules.yaml was changed
+
+Kexa/rules/rulesByProvider/httpSetRules.yaml was changed
+
+Kexa/rules/rulesByProvider/kubernetesSetRules.yaml was changed
+
+Kexa/rules/rulesByProvider/o365SetRules.yaml was changed
+
+initKexa.ps1 was changed
+
+Kexa/services/analyse.service.ts was changed
+
+README.md was changed
+
+config/default.json was changed
+
+documentation/Documentation-Kexa.md was changed
+
+
+## 1.8.0-SNAPSHOT.111.ef2090d
+
+### Files added: 0
+
+### Files changed: 1
+
+initKexa.ps1 was changed
+
+